shared: replace UserDTO/AddressDTO with contracts UserResponseDTO, remove Address model

+++ shared/src/commonMain/kotlin/com/mk/kmpshowcase/domain/repository/UserRepository.kt
@@ -4,7 +4,7 @@ import com.mk.kmpshowcase.domain.model.User
 
 interface UserRepository {
 
-    suspend fun getUser(id: Int): User
+    suspend fun getUser(id: Long): User
 
     suspend fun getUsers(): List<User>
